renames QuantileDistributionComparison.svelte to DistributionComparison.svelte
